Refactor SettingsTableViewController with deuque reusable cell for item at IndexPath.
