+++ Client/Frontend/Theme/ThemedTableSectionHeaderFooterView.swift
